feat: adds metadata to operation result (#25)
https://make.atlassian.net/browse/CDM-13775

We want to allow the apps to pass some metadata about the action. For
now it is just a link, but in the future we see a potential to add extra
information.

Since the link needs to be constructed on the FE, we are only describing
a resource to which the FE should point to. The resource ID the main
thing that you point to, however the route could have some other parts
(like team id, scenario id etc.) which would be indicated in the
additionalParams (for example `{ scenarioId: 1, teamId: 2}`)

Both resource ID and the additionalParams are optional, since you could
also be pointing to a static route

![Screenshot 2025-03-24 at 15 23
50](https://github.com/user-attachments/assets/b9f53e8c-ad37-4a2d-a1d9-dfa6ddb47e91)

---------

Co-authored-by: Martin Vokřínek <m.vokrinek@make.com>