ci: enable sonarqube (#51)
Adds SonarQube integration to the project's CI pipeline using CircleCI.
SonarQube will now automatically scan the codebase for quality and
security issues as part of the workflow, and a configuration file has
been added to define the project's analysis settings.

SonarQube may work as a quality gate of PR. This is recommended but not
required.

Part of https://make.atlassian.net/browse/DP-176

---
[Has AI Code]